Revert "Added journal map for new entires, made scroll for new trips because …"
